Adds table-open-cache charm config option, as per lp1402586.
Also increases the default for table_open_cache to 2048, as the default
of 512 is a bit too low (and there is no repercussions with using larger
table cache).

Note: table_cache got renamed to table_open_cache in MySQL 5.1.3, but 
mysql-common didn't address that. Change is made in the charm now so
that future upgrades (from 5.5 to 5.6) can be smooth.