Wait until clustered before running client hooks
Percona cluster takes some time to fully cluster. The charm was
previously running shared-db-relation-changed hooks whenever they
were queued even if the cluster was not yet complete. This may lead
to split brain scenarios or unexpected behavior.

This change confirms the entire cluster is ready before running
client shared-db-relation-changed hooks.

min-cluster-size can now be used to attempt to guarantee the cluster
is ready with the expected number of nodes. If min-cluster-size is
not set the charm will still determine based on the information
available if all the cluster nodes are ready. Single node
deployments are still possible.

Partial-Bug: #1655417
Change-Id: Ie9deb266a9682e86f3a9cbc1103b655b13a8295e