Recheck sync status when assessing status of unit
After a resume operation, or after a boot following a prolonged
period of downtime, it may take some time for the local unit
to re-sync with its peers.

Update the function that assesses local unit status to recheck
sync status up to 10 times (with an increasing delay between
recheck) before declaring that the unit is in a blocked state.

Change-Id: Idaa960ade4c52c9ebba6c65a55bade6b22e90cdc
Closes-Bug: 1626450