Fix differing bootstrap uuids
When is_bootstrapped is consulted it was returning a false negative.
When the leader node is bootstrapped it sets the uuid via leader_set,
but during a leader-settings-changed hook when the non-leader should
be picking up the uuid it was checking relation data instead which will
be way behind the curve. This is a vestigial block of code
pre-leadership.

Leader settings get set much earlier than relation data. This change
consults leader_get rather than the relation. We also make sure the
mysqld.cnf file is rendered on a leader-settings-changed hook.

Change-Id: I95e56bd28152c934f413025a22dd6821b2ad8e94
Closes-Bug: #1738896