Guarantee timing of installation and render
Previous attempts to solve Bug #1738896 missed the root cause. The
root cause problem is when the configuration file is rendered before
percona is installed. The rendering includes clustering configuration
which causes percona-cluster to automatically do a single bootstrap
when percona-cluster packages are installed leading to the UUID
mismatch.

The timing and ordering of installation, rendering of the
configuration and restart of mysql is critical across all hook
executions.

This change is a partial reversion of Change ID
I95e56bd28152c934f413025a22dd6821b2ad8e94. The change primarily
guarantees percona-cluster is not installed on non-leader nodes
before the leader is bootstrapped and makes sure the configuration
does not get rendered prior to installation.

is_leader_bootstrapped is introduced and guarantees all data expected
from the leader is available to guard on various tasks.

Closes-Bug: #1744961
Closes-Bug: #1738896
Change-Id: Ifeb1520dba3b14fc1b51a586141905a385f2b2c1