Retrieve root password from leaderdb if needed
Within update_root_password if there is no root password in the
config() object then cfg['root-password'] will return None.
set_mysql_root_password is then called with None which triggers a
TypeError. The desired root password is in the leaderdb so fallback
to retrieving it from there if it is not in config().

unit_tests/test_utils.py was also updated to fix the behaviour of
config.changed to mimic that of the real method in charm helpers.

Change-Id: I8d22a66f335c0c8e893cda699a55056476e6d9d5