Gate db{-admin} relations until cluster is ready
Percona-cluster was responding to db and db-admin relations before it
was ready. This led to the error: "WSREP has not yet prepared node for
application use."

This change applies the same gating share-db relation already has to db
and db-admin relations. It also condenses code used in both instances.

This change guarantees the rendered configuration will not
auto-bootstrap for non-leaders. This addresses Bug 1738896.

Closes-Bug: #1742683
Closes-Bug: #1738896
Change-Id: If525595fd109e6a738071a3f016b9c2eabec529e