Ensure leader settings on charm upgrade
Currently bootstrapping is gated by is_leader_bootstrapped which
checks a handful of leader settings. When upgrading from older
versions of the charm, these settings are missing leading to an
attempt to bootstrap an already bootstrapped cluster.

This change makes sure the leader settings is_leader_bootstrapped is
checking for are all set by the leader on upgrade-charm.

Closes-Bug: #1755507

Change-Id: I172f10221b9447ca3e0c5403feaa49acccfa9e42