Make pxc-strict-mode configurable via charm
PXC5.7, which is default one on bionic, introduces the new config
option, pxc_stric_mode, which is set to 'enforcing' by default.

This, however, prevents installing older versions of Openstack
with PXC5.7 on bionic (use case is, of course, testing).

This change makes pxc-strict-mode configurable via the charm, and
it defaults it to 'enforcing', but user can change it at hers or
his desire.

This option is going to be ignored on PXC installed on series less
than bionic.

Change-Id: I7d82e8bea2e75da6159b5baf2e88f2bd44ce84b3